feat(web): add Clients page
