+++ glados-web/templates/base.html
@@ -33,6 +33,9 @@
                 <li class="nav-item">
                     <a class="nav-link active" aria-current="page" style="margin-left: 20px;" href="/census/explorer">Census Explorer</a>
                 </li>
+                <li class="nav-item">
+                    <a class="nav-link active" aria-current="page" style="margin-left: 20px;" href="/clients">Clients</a>
+                </li>
             </ul>
             <select name="network-selector" id="network-selector" class="form-select" style="width: auto;">
                 <option value="History">History</option>
