+++ glados-web/templates/index.html
@@ -164,7 +164,7 @@
                     </div>
                     <div class="table-responsive">
                         <select
-                            id="audit-block-status-content-type"
+                            id="audit-block-status-select"
                             class="form-select"
                             style="width: auto"
                         >
@@ -190,7 +190,9 @@
                         This graph shows the amount of nodes that were reachable at each census (every 15 minutes).
                     </div>
                     <div class="table-responsive">
-                        <div id="census-weekly-graph"></div>
+                        <div id="census-weekly-graph-container">
+                            <div id="census-weekly-graph"></div>
+                        </div>
                     </div>
                 </div>
             </div>
@@ -206,11 +208,61 @@
 
 </script>
 <script type="module">
-    import {censusWeeklyChart} from "../static/js/census_weekly.js";
-    censusWeeklyChart();
+    import {areaGraph} from "../static/js/area_graph.js";
 
-    import {auditBlockStatusChart} from "../static/js/audit_block_status.js";
-    auditBlockStatusChart();
+
+  {% if subprotocol == SubProtocol::History %}
+    areaGraph({
+      baseName: "audit-block-status",
+      baseUrl: "/api/audit-block-status/?",
+      graphTitle: "Status of latest 4444 audit per block",
+      labelY: "% of blocks",
+      kind: "zoom",
+      dataShape: "wide",
+      valueAttribute: "nodeCount",
+      stackHundrethPercent: true,
+      selectParam: "content_type",
+      seriesMetadata: [
+        {
+          slug: "success",
+          color: "#90ee90",
+          name: "Success",
+        },
+        {
+          slug: "error",
+          color: "#ffc0cb",
+          name: "Error",
+        },
+        {
+          slug: "unaudited",
+          color: "#dddddd",
+          name: "Unaudited"
+        },
+      ],
+      initialRanges: {
+        x: [
+             new Date(Date.UTC(2015, 6, 30, 15, 26, 13)), // GENESIS_TIMESTAMP: 2015-07-30T15:26:13Z
+             new Date(Date.UTC(2022, 8, 15, 6, 42, 42)),  // LAST_MINED_BLOCK_TIMESTAMP: 2022-09-15T06:42:42Z
+        ],
+        y: [0, 1],
+      }
+  });
+  {% endif %}
+
+    areaGraph({
+      baseName: "census-weekly",
+      baseUrl: "/api/census-weekly/?",
+      graphTitle: "Nodes in Census, by week",
+      labelY: "# of nodes",
+      kind: "weekly",
+      dataShape: "singleSeries",
+      color: "lightblue",
+      valueAttribute: "nodeCount",
+      initialRanges: {
+        x: [Date.now() - 7 * 24 * 60 * 60 * 1000, Date.now()],
+        y: [0, 100],
+      }
+    });
 </script>
 
 {% endblock %}
